+++ component/include/pl_blackbox_configuration.h
@@ -11,6 +11,8 @@ class BlackBoxConfiguration {
 public:
   BlackBoxConfiguration() = default;
   virtual ~BlackBoxConfiguration() = default;
+  BlackBoxConfiguration(const BlackBoxConfiguration&) = delete;
+  BlackBoxConfiguration& operator=(const BlackBoxConfiguration&) = delete;
 
   /// @brief Loads the configuration
   virtual void Load() = 0;